feat: Implement complete backend and refactor data types
- Implement all layers: Model, Repository, Service, and Controller.
- Refactor price field from double to BigDecimal for precision.
- Establish JPA relationships for all entities.
- Implement DTO pattern for all API endpoints.
- Expose full CRUD for Products and Clients.
- Add core logic for Shopping Cart management.

+++ src/main/java/br/com/thalesos/supermarket_api/dto/ProductRequestDTO.java
@@ -1,4 +1,6 @@
 package br.com.thalesos.supermarket_api.dto;
 
-public record ProductRequestDTO(String name, double price) {
+import java.math.BigDecimal;
+
+public record ProductRequestDTO(String name, BigDecimal price) {
 }
